chore(tactile): align style with the joint side of the package
The tactile and joint driver share `wujihand_driver` as their package and
namespace, but the tactile files landed with style choices that diverged
from the joint side that has been in main for a while. Bringing the two
in line so the package looks like it was written by one team:

* Apache 2.0 license headers added to all four tactile files (the joint
  side has them on every cpp/hpp).
* Indentation switched from 4-space to 2-space (Google C++ style, what
  the joint side uses).
* Header guards: `#pragma once` → `#ifndef WUJIHAND_DRIVER__X_HPP_`
  with matching `#endif` comment, matching `wujihand_driver_node.hpp`.
* Visibility labels (`public:` / `private:`) now use the same 1-space
  indent the joint header uses.
* `tactile_main.cpp` renamed to `tactile_driver_main.cpp` so the entry
  point file name follows the same `<thing>_driver_*.cpp` shape as
  `tactile_driver_node.cpp` (the joint-side entry stays at `main.cpp`
  to avoid widening this PR; future cleanup can rename it for full
  symmetry without breaking anything).
* In-source comment that mentioned `tactile_main.cpp` updated to the
  new file name.

No behavioral changes — only formatting, file rename, and one CMake
source-list edit. HIL-verified: tactile_driver_node still streams at
117 Hz, diagnostics at 10 Hz, set_tactile_sample_rate returns success.
