+++ wujihand_driver/include/wujihand_driver/colormap.hpp
@@ -1,4 +1,19 @@
-#pragma once
+// Copyright 2025 Wuji Robotics
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef WUJIHAND_DRIVER__COLORMAP_HPP_
+#define WUJIHAND_DRIVER__COLORMAP_HPP_
 
 #include <cstdint>
 
@@ -45,3 +60,5 @@ inline constexpr uint8_t JET_LUT[256][3] = {
 // clang-format on
 
 }  // namespace wujihand_driver
+
+#endif  // WUJIHAND_DRIVER__COLORMAP_HPP_
